pyproject.toml
[project]
name = "business-workflow-agent"
version = "0.1.0"
description = "Multi-step business workflow agent with tool integration, human-in-the-loop approval, and fine-tuning evaluation"
requires-python = ">=3.11,<3.14"
dependencies = [
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-community>=0.3.0",
    "chromadb>=0.5.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "pydantic>=2.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0",
]

[project.optional-dependencies]
fine-tuning = [
    "torch>=2.2.0",
    "transformers>=4.44.0",
    "peft>=0.13.0",
    "datasets>=3.0.0",
    "accelerate>=1.0.0",
    "bitsandbytes>=0.44.0",
    "trl>=0.12.0",
    "scikit-learn>=1.5.0",
    "matplotlib>=3.9.0",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.7.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"
